fix: 🐛 調整登入驗證
在 (auth) 下新增統一驗證，未登入即導回首頁、點擊立即報名跳出登入彈窗
