+++ src/app/(auth)/events/[eventId]/checkout/page.tsx
@@ -1,21 +1,8 @@
 "use client";
 import { useAuthStore } from "@/store/auth";
-import { useDialogStore } from "@/store/dialog";
-import { useEffect } from "react";
-import { toast } from "sonner";
 
 export default function CheckoutPage() {
   const isAuthenticated = useAuthStore((s) => s.isAuthenticated);
-  const setLoginDialogOpen = useDialogStore((s) => s.setLoginDialogOpen);
-  const setLoginTab = useDialogStore((s) => s.setLoginTab);
-
-  useEffect(() => {
-    if (!isAuthenticated) {
-      setLoginTab("signin");
-      setLoginDialogOpen(true);
-      toast.error("請先登入才能購票");
-    }
-  }, [isAuthenticated, setLoginDialogOpen, setLoginTab]);
 
   if (!isAuthenticated) return null;
