+++ src/app/(auth)/layout.tsx
@@ -1,0 +1,17 @@
+"use client";
+import { useAuthStore } from "@/store/auth";
+import { useRouter } from "next/navigation";
+import { useEffect } from "react";
+
+export default function AuthLayout({ children }: { children: React.ReactNode }) {
+  const isAuthenticated = useAuthStore((s) => s.isAuthenticated);
+  const router = useRouter();
+
+  useEffect(() => {
+    if (!isAuthenticated) {
+      router.replace("/");
+    }
+  }, [isAuthenticated, router]);
+  if (!isAuthenticated) return null;
+  return <>{children}</>;
+}
